+++ changelog.md
@@ -1,5 +1,3 @@
 ### Changelog
 
-* Added support for more library and plugin notations in catalog.
-* Repositories configuration file replaced with kataloger configuration file, where beside repositories can be specified catalog paths and parameters.
-* Minimum required python version lowered to 3.8.
+* Fix: Default configuration file now correctly shipped.